refactor(coordinator): create TickScheduler skeleton
- Add TickScheduler class with basic structure
- No behavior change - skeleton only
- Part of coordinator refactoring (#751)
