+++ custom_components/localshift/coordinator/tick_scheduler.py
@@ -1,0 +1,38 @@
+"""Periodic task scheduling for LocalShift coordinator.
+
+Responsibilities:
+- FAST tick (1 min): state machine evaluation, automation readiness
+- MEDIUM tick (5 min): entity health, learning tasks, load refresh
+- SLOW tick (30 min): weather forecast, forecast accuracy
+- Daily events: midnight reset, daily summary
+- Solar backfill tracking
+- Cost accumulation
+"""
+
+from __future__ import annotations
+
+import logging
+from datetime import datetime
+from typing import TYPE_CHECKING
+
+from homeassistant.core import Event, callback
+
+if TYPE_CHECKING:  # pragma: no cover
+    from .coordinator import LocalShiftCoordinator
+
+_LOGGER = logging.getLogger(__name__)
+
+
+class TickScheduler:
+    """Manages periodic task execution for coordinator."""
+
+    def __init__(
+        self,
+        coordinator: LocalShiftCoordinator,
+    ) -> None:
+        """Initialize tick scheduler.
+
+        Args:
+            coordinator: Parent coordinator instance
+        """
+        self._coordinator = coordinator
