+++ tests/coordinator/test_coordinator.py
@@ -10,25 +10,6 @@ from custom_components.localshift.coordinator import LocalShiftCoordinator
 
 
 @pytest.fixture
-def mock_hass_with_services():
-    """Create a mock Home Assistant instance with services."""
-    hass = MagicMock()
-    hass.services = MagicMock()
-    hass.services.async_call = AsyncMock()
-    hass.states = MagicMock()
-    # async_create_task receives coroutines - use a mock that consumes them
-    # to avoid "coroutine was never awaited" warnings
-    hass.async_create_task = MagicMock(side_effect=lambda coro, name=None: None)
-    return hass
-
-
-@pytest.fixture
-def coordinator(mock_hass_with_services, mock_entry):
-    """Create a LocalShiftCoordinator instance."""
-    return LocalShiftCoordinator(mock_hass_with_services, mock_entry)
-
-
-@pytest.fixture
 def coordinator_data():
     """Create basic CoordinatorData for coordinator tests."""
     from custom_components.localshift.coordinator import CoordinatorData
